Fix: MissingFieldAndHourType

Approved-by: Siwach Toprasert

+++ Dtos/StudyCourseDtos/StaffCoursesDetailResponseDto.cs
@@ -17,7 +17,7 @@ namespace griffined_api.Dtos.StudyCourseDtos
         public Method Method { get; set; }
         public string StartDate { get; set; } = string.Empty;
         public string EndDate { get; set; } = string.Empty;
-        public int TotalHour { get; set; }
+        public double TotalHour { get; set; }
         public int HourPerDay { get; set; }
         public List<ScheduleResponseDto> Schedules { get; set; } = new();
     }
