+++ Models/StudyCourse.cs
@@ -16,7 +16,7 @@ namespace griffined_api.Models
 
         public string Section { get; set; } = string.Empty;
 
-        public int TotalHour { get; set; }
+        public double TotalHour { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
